Move `amount_of_time_formatter()` to `cogs/stats.py`
