+++ cogs/stats.py
@@ -22,6 +22,23 @@ if TYPE_CHECKING:
     from matplotlib.text import Text as Plot_Text
 
 
+def amount_of_time_formatter(value: float, time_scale: str) -> str:
+    """
+    Format the amount of time value according to the provided time_scale.
+
+    E.g. past "1 days" => past "day",
+    past "2.00 weeks" => past "2 weeks",
+    past "3.14159 months" => past "3.14 months"
+    """
+    if value == 1 or float(f"{value:.2f}") == 1:
+        return f"{time_scale}"
+
+    if value % 1 == 0 or float(f"{value:.2f}") % 1 == 0:
+        return f"{int(value)} {time_scale}s"
+
+    return f"{value:.2f} {time_scale}s"
+
+
 def plot_bar_chart(data: dict[str, int], x_label: str, y_label: str, title: str, filename: str, description: str, extra_text: str = "") -> discord.File:  # noqa: E501
     """Generate an image of a plot bar chart from the given data & format variables."""
     plt.style.use("cyberpunk")
@@ -224,7 +241,7 @@ class StatsCommandsCog(TeXBotCog):
                 x_label="Role Name",
                 y_label=(
                     f"""Number of Messages Sent (in the past {
-                        utils.amount_of_time_formatter(
+                        amount_of_time_formatter(
                             settings["STATISTICS_DAYS"].days,
                             "day"
                         )
@@ -333,7 +350,7 @@ class StatsCommandsCog(TeXBotCog):
                     x_label="Role Name",
                     y_label=(
                         f"""Number of Messages Sent (in the past {
-                        utils.amount_of_time_formatter(
+                        amount_of_time_formatter(
                             settings["STATISTICS_DAYS"].days,
                             "day"
                         )
@@ -356,7 +373,7 @@ class StatsCommandsCog(TeXBotCog):
                     x_label="Channel Name",
                     y_label=(
                         f"""Number of Messages Sent (in the past {
-                            utils.amount_of_time_formatter(
+                            amount_of_time_formatter(
                                 settings["STATISTICS_DAYS"].days,
                                 "day"
                             )
@@ -440,7 +457,7 @@ class StatsCommandsCog(TeXBotCog):
                 x_label="Channel Name",
                 y_label=(
                     f"""Number of Messages Sent (in the past {
-                        utils.amount_of_time_formatter(
+                        amount_of_time_formatter(
                             settings["STATISTICS_DAYS"].days,
                             "day"
                         )
